Move binonces to separate file
